requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
requests>=2.31.0

# Data Processing
scanpy>=1.9.3
pandas>=2.1.4
numpy>=1.26.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# LLM Integration (vendor-agnostic)
openai>=1.3.5
anthropic>=0.7.7
huggingface-hub>=0.19.4
transformers>=4.35.2

# Clustering Dependencies (for Scanpy)
scikit-misc>=0.1.0
python-igraph>=0.11.0
leidenalg>=0.10.0

# File Handling
aiofiles>=23.2.1

# Utilities
typing-extensions>=4.8.0

# Development & Testing
black>=23.12.0
flake8>=6.1.0
pytest>=7.4.3
pytest-cov>=4.1.0
mypy>=1.7.0